Split map component to contain the side panel as well
