+++ src/app/map/esri-map/esri-map.component.html
@@ -1,0 +1,1 @@
+<div id="viewDiv"></div>
