+++ src/app/map/esri-map/esri-map.component.scss
@@ -1,0 +1,12 @@
+#viewDiv {
+  position: absolute;
+  width: 100%;
+  top: 64px;
+  bottom: 0;
+}
+
+@media (max-width: 600px) {
+  #viewDiv {
+    top: 56px;
+  }
+}
