+++ src/app/map/map.component.scss
@@ -1,12 +1,6 @@
-#viewDiv {
-  position: absolute;
-  width: 100%;
-  top: 64px;
-  bottom: 0;
-}
 
-@media (max-width: 600px) {
-  #viewDiv {
-    top: 56px;
-  }
+.btn-add-tree {
+  position: absolute !important;
+  bottom: 25px;
+  left: 10px;
 }
