+++ src/app/map/side-panel/side-panel.component.scss
@@ -1,0 +1,17 @@
+.side-panel {
+  position: absolute;
+  top: 64px;
+  right: 0;
+  bottom: 0;
+  width: 30%;
+  min-width: 370px;
+  padding: 10px;
+  box-sizing: border-box;
+  background-color: rgba(255, 255, 255, 0.95);
+
+  .btn-edit {
+    position: absolute;
+    top: 45px;
+    right: 10px;
+  }
+}
