+++ src/app/members/login/login.component.scss
@@ -11,14 +11,14 @@
   margin-top: 30px;
   margin-left:-100px;
   z-index: 10;
-}
 
-button {
-  padding:0.5em;
-  width:100%;
-  cursor:pointer;
-  margin-bottom:12px;
-  font-size:0.7em;
+  button {
+    padding:0.5em;
+    width:100%;
+    cursor:pointer;
+    margin-bottom:12px;
+    font-size:0.7em;
+  }
 }
 
 .basic-btn {
